refactor: break down fixers module into smaller
